Ok there's a lot going on here.
couldnt fix the database problem. please talk to mo about that in class i left a note in the main page where we can view student info. dont mind the dummy data i was working on something that can act in place so were at least part way trhough it